Rework pass structure into clear fwd/rev split
