--- mlir/lib/Quantum/Transforms/AdjointLowering/adjoint_lowering.cpp
@@ -1,66 +1,0 @@
-// Copyright 2023 Xanadu Quantum Technologies Inc.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-
-//     http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#define DEBUG_TYPE "adjoint"
-
-#include <memory>
-#include <vector>
-
-#include "llvm/Support/Debug.h"
-#include "llvm/Support/Errc.h"
-#include "mlir/Dialect/Arith/IR/Arith.h"
-#include "mlir/Dialect/Bufferization/IR/Bufferization.h"
-#include "mlir/Dialect/Complex/IR/Complex.h"
-#include "mlir/Dialect/Index/IR/IndexDialect.h"
-#include "mlir/Dialect/Linalg/IR/Linalg.h"
-#include "mlir/Dialect/MemRef/IR/MemRef.h"
-#include "mlir/Dialect/SCF/IR/SCF.h"
-#include "mlir/Dialect/Tensor/IR/Tensor.h"
-#include "mlir/IR/BuiltinOps.h"
-#include "mlir/IR/SymbolTable.h"
-#include "mlir/Pass/Pass.h"
-#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
-
-#include "Catalyst/IR/CatalystDialect.h"
-#include "Quantum/IR/QuantumOps.h"
-#include "Quantum/Transforms/Patterns.h"
-
-using namespace llvm;
-using namespace mlir;
-using namespace catalyst::quantum;
-
-namespace catalyst {
-namespace quantum {
-
-#define GEN_PASS_DEF_ADJOINTLOWERINGPASS
-#include "Quantum/Transforms/Passes.h.inc"
-
-struct AdjointLoweringPass : impl::AdjointLoweringPassBase<AdjointLoweringPass> {
-    using AdjointLoweringPassBase::AdjointLoweringPassBase;
-
-    void runOnOperation() final
-    {
-        LLVM_DEBUG(dbgs() << "adjoint lowering pass"
-                          << "\n");
-
-        RewritePatternSet patterns(&getContext());
-        populateAdjointPatterns(patterns);
-        if (failed(applyPatternsGreedily(getOperation(), std::move(patterns)))) {
-            return signalPassFailure();
-        }
-    }
-};
-
-} // namespace quantum
-} // namespace catalyst
